requirements.txt
pandas~=1.2.5
PyYAML~=5.4.1